+++ .babelrc
@@ -2,6 +2,7 @@
   "presets": [ "es2015" ],
   "plugins": [
     "add-module-exports",
-    "transform-es2015-modules-umd"
+    "transform-es2015-modules-umd",
+    "transform-object-rest-spread"
   ]
 }
