+++ package.json
@@ -56,6 +56,7 @@
     "babel-loader": "^6.2.5",
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-transform-es2015-modules-simple-amd": "^0.3.0",
+    "babel-plugin-transform-object-rest-spread": "^6.8.0",
     "babel-preset-es2015": "^6.14.0",
     "eslint": "^3.4.0",
     "eslint-loader": "^1.5.0",
