+++ src/chaining/index.js
@@ -1,0 +1,18 @@
+/**
+ * Created by Paul on 9/7/2016.
+ */
+
+import opt from './opt';
+import chain from './chain';
+import put from './put';
+import get from './get';
+import key from './key';
+import on from './on';
+import path from './path';
+import map from './map';
+import val from './val';
+import not from './not';
+import set from './set';
+import init from './init';
+
+export default {opt, chain, put, get, key, on, path, map, val, not, set, init};
